refactor(storage): keep only what the two extractors actually spell differently
`ZipExtractor` and `TarExtractor` differed in opening the archive and listing its members;
closing and extracting were identical. Those two move up into `BaseExtractor`, which is now
generic over the archive type so `self.archive` stays concrete in each subclass.
